moving code that access to storage from controller to services
